requirements.txt
# numerical computing
numpy==1.21.0

# TF variants 
pydot-ng==2.0.0
tensorflow_datasets==4.1.0
tensorflow_addons==0.16.1
tensorflow-hub==0.12.0
#tf-models-official==2.9.2

# image processing
scikit-image==0.17.2
opencv-python==4.4.0.46

# data science
pandas==1.4.2
scikit-learn==1.0.2

# visualizer
matplotlib==3.3.3
seaborn==0.11.2

# utils
tqdm==4.54.0
easydict==1.9
jupyter==1.0.0
imgcat==0.5.0